Add files via upload
Manipulates characters in a string based off their positions in the alphabet. Replaces every letter in the string with the letter following it in the alphabet (i.e. c becomes d, z becomes a). Then capitalize every vowel in this new string (a, e, i, o, u) and finally returns this modified string. Written for a Coderbyte coding challenge.