Add files via upload
This Python script computes the average score (or any value) for a student (or any object). The first input is the number of students (or objects). Next, the script takes the name of each student followed by their scores on separate lines. Finally, the user inputs a students' name and their average score is returned. Completed for a coding challenge.